backend/requirements.txt
fastapi
uvicorn[standard]
groq
tavily-python
trafilatura
sentence-transformers
torch
python-dotenv
requests
sse-starlette
scipy
numpy
